support third-party plugin

+++ client/lib/plugin_builder.Dockerfile
@@ -6,31 +6,34 @@ RUN apt-get install -qy zip nodejs build-essential \
   binutils-aarch64-linux-gnu gcc-aarch64-linux-gnu  g++-aarch64-linux-gnu
 RUN npm install -g node-gyp yarn
 
-CMD cp -r /plugin /build && \
+CMD \
+    cp -r /plugin /build && \
     cd /build && \
-    yarn && \
-    \
-    mkdir -p native/x64 && \
-    mv build/Release/*.node native/x64 && \
-    rm -rf build && \
-    \
-    AR=arm-linux-gnueabihf-ar \
-    CC=arm-linux-gnueabihf-gcc \
-    CXX=arm-linux-gnueabihf-g++ \
-    LINK=arm-linux-gnueabihf-g++ \
-    node-gyp rebuild --arch arm && \
-    mkdir -p native/arm && \
-    mv build/Release/*.node native/arm && \
-    rm -rf build && \
-    \
-    AR=aarch64-linux-gnu-ar \
-    CC=aarch64-linux-gnu-gcc \
-    CXX=aarch64-linux-gnu-g++ \
-    LINK=aarch64-linux-gnu-g++ \
-    node-gyp rebuild --arch arm64 && \
-    mkdir -p native/arm64 && \
-    mv build/Release/*.node native/arm64 && \
-    rm -rf build && \
-    \
-    rm -rf node_modules/nan yarn.lock binding.gyp && \
+    if [ -f package.json ]; then \
+      yarn && \
+      \
+      mkdir -p native/x64 && \
+      mv build/Release/*.node native/x64 && \
+      rm -rf build && \
+      \
+      AR=arm-linux-gnueabihf-ar \
+      CC=arm-linux-gnueabihf-gcc \
+      CXX=arm-linux-gnueabihf-g++ \
+      LINK=arm-linux-gnueabihf-g++ \
+      node-gyp rebuild --arch arm && \
+      mkdir -p native/arm && \
+      mv build/Release/*.node native/arm && \
+      rm -rf build && \
+      \
+      AR=aarch64-linux-gnu-ar \
+      CC=aarch64-linux-gnu-gcc \
+      CXX=aarch64-linux-gnu-g++ \
+      LINK=aarch64-linux-gnu-g++ \
+      node-gyp rebuild --arch arm64 && \
+      mkdir -p native/arm64 && \
+      mv build/Release/*.node native/arm64 && \
+      rm -rf build && \
+      \
+      rm -rf node_modules/nan yarn.lock binding.gyp; \
+    fi; \
     zip -FSr /dist/plugin.zip *
